+++ realm/realm-library/src/main/java/io/realm/BaseRealm.java
@@ -340,8 +340,12 @@ abstract class BaseRealm implements Closeable {
      * incompatible schema changes.
      */
     public void beginTransaction() {
+        beginTransaction(false);
+    }
+
+    void beginTransaction(boolean ignoreReadOnly) {
         checkIfValid();
-        sharedRealm.beginTransaction();
+        sharedRealm.beginTransaction(ignoreReadOnly);
     }
 
     /**